:recycle: refactor: ObjectMapper를 주입받아 사용하도록 변경
